Miscellaneous changes
- minor mods to hybrid_tests
- fixes to PAPI components
- PAPI components were separated into own files
- PAPI tests no longer pass with zeros
- Added papi_rate_tuple tests
- PAPI rate tuple template signature changed, first param is now the component divisor
- fixed issues with gotcha bindings in MT
- removed the default component_bundle initializer
- compiler instrumentation is now more efficient
- timemory-run no longer regexes internal file names
